% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artificial_signal.R
\name{artificial_signal}
\alias{artificial_signal}
\title{Artificial signal}
\usage{
artificial_signal(
  duration,
  uncorr_rate = 0,
  hists_rate = 0,
  fission_multiplicity = NULL,
  k = 0,
  lambda = 1,
  hists_id = FALSE
)
}
\arguments{
\item{duration}{Signal duration (unitless).}

\item{uncorr_rate}{Number of uncorrelated detections per time unite.}

\item{hists_rate}{Number of histories per time unite. Every creted neutron create a detection.}

\item{fission_multiplicity}{Same argument as \code{\link[NeutronNoise]{get_fission_chains_length}}.}

\item{k}{Same argument as \code{\link[NeutronNoise]{get_fission_chains_length}}.}

\item{lambda}{Exponential decay parameter of each history. Parameter rate of \code{\link[stats]{rexp}}.}

\item{hists_id}{If TRUE also generate virtual history identifiers.}
}
\value{
A sorted data.table containing at lease a TIME column.
}
\description{
Creates an artificial signal (i.e. not from sensors or simulations) 
with uncorrelated and/or correlated components.
Delayed neutrons are not take into account.
Signal start from time 0.
}
\examples{
artificial_signal(duration = 10, uncorr_rate = 5, hists_rate = 20, 
fission_multiplicity = data.frame(nu=1:2,pdf=c(0.5,0.5)), k = 0.9, lambda = 5, hists_id  = TRUE)

}
