% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_signal.R
\name{as_signal.numeric}
\alias{as_signal.numeric}
\title{Numeric to signal}
\usage{
\method{as_signal}{numeric}(x, x_duration, ...)
}
\arguments{
\item{x}{Numeric vector containting time detections.}

\item{x_duration}{Signal duration (unitless).}

\item{...}{Unused.}
}
\value{
A sorted data.table containing a TIME column.
}
\description{
Wrapper to create a signal from a numeric vector.
}
\examples{
as_signal(runif(200, 0, 20), x_duration = 20) \%>\% plot()

}
