% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_signal.R
\name{as_signal}
\alias{as_signal}
\title{Generic to signal}
\usage{
as_signal(x, x_duration, ...)
}
\arguments{
\item{x}{Data containting time detections.}

\item{x_duration}{Signal duration (unitless).}

\item{...}{Unused.}
}
\value{
A sorted data.table containing a TIME column.
}
\description{
Wrapper to create a signal from various data type.
See \code{\link[NeutronNoise]{as_signal.numeric}} and 
\code{\link[NeutronNoise]{as_signal.data.frame}} for specialised version.
}
