% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useful.R
\name{lseq}
\alias{lseq}
\title{Logarithmic sequence}
\usage{
lseq(from, to, length)
}
\arguments{
\item{from}{Lower limit.}

\item{to}{Upper limit.}

\item{length}{Sequence length.}
}
\value{
Numeric vector
}
\description{
Code taken from the "emdbook" package. Create a logarithmic sequence between \code{from} and \code{to} of size \code{length}.
}
\examples{
lseq(1, 10, 7)

}
