% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feynman_v2m_plot.R
\name{plot.feynman_v2m}
\alias{plot.feynman_v2m}
\title{Plot Feynman Variance-to-Mean curve}
\usage{
\method{plot}{feynman_v2m}(x, ...)
}
\arguments{
\item{x}{data.frame from \link[NeutronNoise]{feynman_v2m} functions.}

\item{...}{Unused.}
}
\value{
ggplot object.
}
\description{
Plot Feynman Variance-to-Mean curve from \code{\link[NeutronNoise]{feynman_v2m}} function. 
The assumed time unit is second.
}
\examples{
mls <- data.frame(nu = 0:4, pdf = c(0.1,0.3,0.35,0.4,0.2))
artificial_signal(1000, hists_rate = 5000, 
                  fission_multiplicity = mls, k = 0.9, lambda = 10) \%>\%
feynman_v2m(samples_widths = lseq(from = 0.001, to = 10, 50)) \%>\% plot()

}
