% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_plot.R
\name{plot.signal}
\alias{plot.signal}
\title{Plot a signal}
\usage{
\method{plot}{signal}(x, binwidth = 1, ...)
}
\arguments{
\item{x}{A data.frame from signal function like \link[NeutronNoise]{artificial_signal}.}

\item{binwidth}{Same argument as \link[ggplot2]{geom_histogram}.}

\item{...}{Unused.}
}
\value{
ggplot object.
}
\description{
Plot a signal. The assumed time unit is second.
}
\examples{
artificial_signal(duration = 1000, uncorr_rate = 5) \%>\% plot()

}
