% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcr.R
\name{xcr}
\alias{xcr}
\title{Calculate single counte rate, double counte rate and triple counte rate}
\usage{
xcr(x)
}
\arguments{
\item{x}{feynman_hist object from \link[NeutronNoise]{feynman_hist} functions.}
}
\value{
xcr object which is a tibble object.
}
\description{
Thoses count rates are calculated base on the feynman histograms obtain from 
 \code{\link[NeutronNoise]{feynman_hist}} function. 
The assumed time unit is second.

The calculated values are:

-  Y1: single count rate
-  Y2: double count rate
-  Y3: triple count rate
-  Y1_std: standard deviation of Y1
-  Y2_std: standard deviation of Y2
-  Y3_std: standard deviation of Y3
-  Y: Feynman value
-  Y_std: standard deviation of Y
}
\examples{
artificial_signal(1000, 5) \%>\% feynman_hist(samples_widths = c(0.8, 0.9)) \%>\% xcr()

}
