% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_signal.R
\name{as_signal.data.frame}
\alias{as_signal.data.frame}
\title{data.frame to signal}
\usage{
\method{as_signal}{data.frame}(x, x_duration, is_t0 = FALSE, ...)
}
\arguments{
\item{x}{data.frame containing at least a TIME column with numeric values.}

\item{x_duration}{Signal duration (unitless).}

\item{is_t0}{If TRUE this means that all source events were generated at time 0. 
The function then distributes the histories over the specified time interval \code{x_duration}. 
\code{x} must contains in addition to the TIME column the HIST column (story identifier).}

\item{...}{Unused.}
}
\value{
A sorted data.table containing a TIME column.
}
\description{
Wrapper to create a signal from a numeric vector.
}
\examples{
as_signal(data.frame(TIME = runif(200, 0, 20)), x_duration = 20) \%>\% plot()

}
