% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{feynman_hist}
\alias{feynman_hist}
\title{Feynman histogram}
\usage{
feynman_hist(x, samples_widths, max_nb_samples = 0L, verbose = 0L)
}
\arguments{
\item{x}{A sorted numeric vector representing the detection times. The function assumes that the signal starts at the time 0 seconds.}

\item{samples_widths}{Numeric vector of samples width (multiple values of \code{T}).}

\item{max_nb_samples}{If different from 0 then the calculation is limited to the specified number of samples.}

\item{verbose}{For debbuging purpose only.}
}
\value{
A DataFrame.
}
\description{
This function divide the signal into equal samples of width \code{T}. 
For each of them the number of events are counted and binned into a histogram called the Feynman histogram.
This histogram therefore represents the occurrence probabilities of various multiplets (i.e. 1 detection, 2 detections, etc.) occuring within
a specified time gate width \code{T}.
To avoid numerical problem with functions using \code{feynman_hist} the last sample is not taken into account.
}
